# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Music Search Engine API
  description: |
    API for a personal music library web application for uploading, organizing,
    searching, and streaming legal music files.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: https://api.example.com/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
  /me:
    get:
      tags: [User]
      summary: Get current user profile
      operationId: getCurrentUser
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [User]
      summary: Update current user profile
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /upload/presigned:
    post:
      tags: [Upload]
      summary: Get presigned URL for file upload
      operationId: getPresignedUploadURL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresignedUploadRequest'
      responses:
        '200':
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /upload/confirm:
    post:
      tags: [Upload]
      summary: Confirm file upload and trigger processing
      operationId: confirmUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmUploadRequest'
      responses:
        '200':
          description: Upload confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /uploads:
    get:
      tags: [Upload]
      summary: List user's uploads
      operationId: listUploads
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/UploadStatus'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/lastKey'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/sortOrder'
      responses:
        '200':
          description: List of uploads
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    lastKey:
      name: lastKey
      in: query
      schema:
        type: string
    sortBy:
      name: sortBy
      in: query
      schema:
        type: string
    sortOrder:
      name: sortOrder
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PayloadTooLarge:
      description: Payload too large
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnsupportedMediaType:
      description: Unsupported media type
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        lastKey:
          type: string
        nextKey:
          type: string
        totalEstimate:
          type: integer
    UserResponse:
      type: object
      required: [id, email, displayName, createdAt, storageUsed, storageLimit, trackCount, albumCount, playlistCount]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        storageUsed:
          type: integer
          format: int64
        storageLimit:
          type: integer
          format: int64
        trackCount:
          type: integer
        albumCount:
          type: integer
        playlistCount:
          type: integer
    UpdateUserRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 100
        avatarUrl:
          type: string
          format: uri
    UploadStatus:
      type: string
      enum: [PENDING, PROCESSING, COMPLETED, FAILED]
    PresignedUploadRequest:
      type: object
      required: [fileName, fileSize, contentType]
      properties:
        fileName:
          type: string
          minLength: 1
          maxLength: 500
        fileSize:
          type: integer
          format: int64
          minimum: 1
          maximum: 524288000
        contentType:
          type: string
          enum: [audio/mpeg, audio/flac, audio/wav, audio/aac, audio/ogg, audio/x-flac]
    PresignedUploadResponse:
      type: object
      required: [uploadId, uploadUrl, expiresAt, maxFileSize]
      properties:
        uploadId:
          type: string
          format: uuid
        uploadUrl:
          type: string
          format: uri
        fields:
          type: object
          additionalProperties:
            type: string
        expiresAt:
          type: string
          format: date-time
        maxFileSize:
          type: integer
          format: int64
    ConfirmUploadRequest:
      type: object
      required: [uploadId]
      properties:
        uploadId:
          type: string
          format: uuid
    ConfirmUploadResponse:
      type: object
      required: [uploadId, status, message]
      properties:
        uploadId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/UploadStatus'
        message:
          type: string
    UploadResponse:
      type: object
      required: [id, fileName, fileSize, fileSizeStr, contentType, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
        fileSize:
          type: integer
          format: int64
        fileSizeStr:
          type: string
        contentType:
          type: string
        status:
          $ref: '#/components/schemas/UploadStatus'
        errorMsg:
          type: string
        trackId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    UploadListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UploadResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
...